Source: small step towards adding type variables to Source.
I originally didn't want to add polymorphism - or hoped I could avoid it.
However, for functions in the standard library, we want both:
 * type inference
 * polymorphic
Type inference I could probably live without, but some kind of
polymorphism is basically non-negotiable.
